Move data type from mock-server to @/types
