+++ packages/client/src/mocks/mock-server.ts
@@ -1,13 +1,5 @@
 import leaderBoardData from './data/mock-leaderbord-data.json';
-
-export type TLeaderBoardData = {
-    name: string;
-    email: string;
-    login: string;
-    winsCount: number;
-    lostCount: number;
-    score: number;
-};
+import { TLeaderBoardData } from '@/types/data-types';
 
 export default class MockServer {
     protected leaderBoardData: TLeaderBoardData[];
