+++ packages/client/src/types/data-types.ts
@@ -1,0 +1,9 @@
+export type TLeaderBoardData = {
+    name: string;
+    login: string;
+    winsCount: number;
+    lostCount: number;
+    score: number;
+};
+
+export default TLeaderBoardData;
